Better class hierarchy
