+++ bokehjs/src/lib/models/glyphs/webgl/base_marker.ts
@@ -57,45 +57,6 @@ export abstract class BaseMarkerGL extends BaseGLGlyph {
 
   abstract override draw(indices: number[], mainglyph: any, trans: Transform): void
 
-  protected _draw_impl(indices: number[], transform: Transform, main_gl_glyph: BaseMarkerGL, marker_type: MarkerType): void {
-    if (main_gl_glyph.data_changed) {
-      main_gl_glyph._set_data()
-      main_gl_glyph.data_changed = false
-    }
-
-    if (this.visuals_changed) {
-      this._set_visuals()
-      this.visuals_changed = false
-    }
-
-    const nmarkers = main_gl_glyph.nvertices
-
-    if (this._show == null)
-      this._show = new Uint8Buffer(this.regl_wrapper)
-
-    const prev_nmarkers = this._show.length
-    const show_array = this._show.get_sized_array(nmarkers)
-    if (indices.length < nmarkers) {
-      this._show_all = false
-
-      // Reset all show values to zero.
-      for (let i = 0; i < nmarkers; i++)
-        show_array[i] = 0
-
-      // Set show values of markers to render to 255.
-      for (let j = 0; j < indices.length; j++) {
-        show_array[indices[j]] = 255
-      }
-    } else if (!this._show_all || prev_nmarkers != nmarkers) {
-      this._show_all = true
-      for (let i = 0; i < nmarkers; i++)
-        show_array[i] = 255
-    }
-    this._show.update()
-
-    this._draw_one_marker_type(marker_type, transform, main_gl_glyph)
-  }
-
   protected _draw_one_marker_type(marker_type: MarkerType, transform: Transform, main_gl_glyph: BaseMarkerGL): void {
     if (this._have_hatch) {
       const props: MarkerHatchGlyphProps = {
