+++ bokehjs/src/lib/models/glyphs/webgl/index.ts
@@ -2,5 +2,6 @@ export {get_regl} from "./regl_wrap"
 export * from "./base_marker"
 export * from "./line_gl"
 export * from "./quad"
-export * from "./marker"
+export * from "./multi_marker"
 export * from "./rect"
+export * from "./single_marker"
